feat: add common layout types
